[update] merged api's dogs , manageDog and create dog
